+++ src/main/java/blackjack/domain/card/Deck.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import blackjack.domain.card.Card;
