+++ src/main/java/blackjack/domain/card/HandGenerator.java
@@ -1,6 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import blackjack.domain.card.Card;
+import blackjack.domain.card.Deck;
 import blackjack.domain.card.Hand;
 import java.util.ArrayList;
 import java.util.List;
