+++ src/main/java/blackjack/domain/participant/Participant.java
@@ -1,7 +1,7 @@
 package blackjack.domain.participant;
 
-import blackjack.domain.Deck;
-import blackjack.domain.HandGenerator;
+import blackjack.domain.card.Deck;
+import blackjack.domain.card.HandGenerator;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Hand;
 import java.util.List;
