+++ src/main/java/blackjack/domain/participant/Participants.java
@@ -1,6 +1,6 @@
 package blackjack.domain.participant;
 
-import blackjack.domain.HandGenerator;
+import blackjack.domain.card.HandGenerator;
 import java.util.List;
 
 public class Participants {
