+++ src/main/java/blackjack/domain/result/HandResult.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.result;
 
 public enum HandResult {
     WIN("승"),
