+++ src/test/java/blackjack/domain/card/HandGeneratorTest.java
@@ -1,8 +1,10 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.domain.card.Deck;
 import blackjack.domain.card.Hand;
+import blackjack.domain.card.HandGenerator;
 import blackjack.domain.card.Number;
 import blackjack.domain.card.Shape;
 import blackjack.testutil.CustomDeck;
