+++ src/test/java/blackjack/domain/participant/PlayersTest.java
@@ -2,9 +2,9 @@ package blackjack.domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import blackjack.domain.Deck;
-import blackjack.domain.HandGenerator;
-import blackjack.domain.RandomDeck;
+import blackjack.domain.card.Deck;
+import blackjack.domain.card.HandGenerator;
+import blackjack.domain.card.RandomDeck;
 import blackjack.exception.InvalidPlayerCountException;
 import java.util.ArrayList;
 import org.junit.jupiter.api.DisplayName;
