+++ src/test/java/blackjack/domain/result/BlackjackResultTest.java
@@ -1,7 +1,9 @@
-package blackjack.domain;
+package blackjack.domain.result;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.domain.card.Deck;
+import blackjack.domain.card.HandGenerator;
 import blackjack.domain.card.Number;
 import blackjack.domain.participant.Name;
 import blackjack.domain.participant.Player;
