+++ src/test/java/blackjack/testutil/CustomDeck.java
@@ -1,6 +1,6 @@
 package blackjack.testutil;
 
-import blackjack.domain.Deck;
+import blackjack.domain.card.Deck;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Number;
 import blackjack.domain.card.Shape;
